Don't use the user's fraction to decide the visibility of a leash
The caller might pass any fraction to
WindowInsetsAnimationController#setInsetsAndAlpha. The fraction should
not control visibility of the leash.

This CL flags mCancelling when an insets animation is being cancelled
instead of setting mPendingFraction to 1f, and stops using
mPendingFraction to decide the leash visibility if the animation is
ANIMATION_TYPE_USER.

Bug: 376232646
Flag: EXEMPT bugfix
Test: Manual, i.e. verified that IME does not jump hide anymore when
      swiping down and up again on all-apps.
Change-Id: I297758f665460bd6db8ff1da8267e9b9a862b328