Remove manual freezing screen
Currently this is only used when switching to a new user. That was a
workaround of bug 279773661. In the case, the UserSwitchingDialog
looks flickering because the animation incorrectly put the top activity
(setupwizard) at a higher z-order than the overlay window, which causes
  dialog (cover on top) -> activity (animating wrong z-order)
     -> dialog (animation finished) -> activity (dialog dismissed)

The incorrect z-order was fixed by the change
If6c12d4c4ee3b19abbe1fe4307dd4acae35e5cf8 (wm.Transition)
Which makes the transition recognize that the top activity is
a dependent change which shouldn't be put at transition root.

So now UserSwitchingDialog can always stay on top during switching
user, then startFreezingScreen/stopFreezingScreen are no longer needed.

Bug: 389860816
Flag: EXEMPT bugfix
Test: Switch to a new user. The animation appearance is the same as
      before this change.
Change-Id: I73685cbfc1beb1fc4066daa870408a23c731f551