Fix FilterClientCallbackImpl race condition
When changing the tuner source (e.g. recording playback to live
viewing), the FilterClientCallbackImpl will be recreated. When releasing
it, the thread sending filter events may still be running depend on the
HAL implementation. We should synchronous the filter event thread and
release thread to avoid the crash in Tuner framework.

Bug: 365688825
Test: test on cf_x86_tv
Flag: EXEMPT bugfix
Change-Id: I84e1d315730fc904b1083a8291c074f0d9d50195