Explicit registraton to refresh rate
To not break the applications using the older sdks, we are implicitly
subscribing the clients to refresh rate events when the
registerDisplayListener(DisplayListener, Handler) API is called.
However, in other variants on registerDisplayListener, we now expect
users to explicitly subscribe to EVENT_TYPE_DISPLAY_REFRESH_RATE to get
the refresh rate updates.

Bug: 387962667
Flag: EXEMPT trivial change
Test: atest DisplayEventTest
Change-Id: I74610cec499546937442b5e10fdecc385d8b9dca