Launch home using the home intent on exit desktop
In some cases we enter desktop from launcher that's not set to NORMAL
state e.g. ALL APPS state. In those scenarios, upon exiting desktop we
would like land on home page in a NORMAL state. This cl makes sure that
when we are exiting desktop, we launch the home with intent so that
launcher sets its states to NORMAL.

Test: atest DesktopTasksControllerTest
Fix: 341328243
Flag: EXEMPT bug fix

Change-Id: Ie13b14882261563f3876c568f6f3c9dbd145166c