Show dialogs on click for OngoingActivityChip Composable
This CL adds a new ClickBehavior field in OngoingActivityChipModel
to define click actions for chips. The ExpandAction is used to
animate the chip into a dialog or activity when it is clicked.
This CL only implements dialogs for now.

Test: Make sure chips expand into a dialog.
Video provided in b/372657935 comment#16
Bug: b/372657935
Flag: com.android.systemui.status_bar_chips_modernization
Change-Id: I284dab65df30246d884a361f4cbbe94558cc5414