Prevent recomposition of VerticalSpannedGrid
We cached the span lists (instead of creating a new one in each
recomposition), because they are unstable. That way, if the contents
don't change, the VerticalSpannedGrid is not recomposed.

Also, move key into the spanned grid.

Test: manual, changes in surrounding code of SpannedGrid don't recompose
it if the spans don't change
Bug: 389985793
Flag: com.android.systemui.qs_ui_refactor_compose_fragment

Change-Id: Ib3e7c9239f7a35b8396b8e7dca07ba3727a06de3