Fix TF visibility due to other translucent TF in Task
The original TF visibility logic considers a TF to be visibile whenever
there is another translucent TF above in the same Task. This seems
incorrect when TFs are stacked, e.g. A|B -> B|C (C is translucent and A
is behind B, but A is determined as visible)

Bug: 392991242
Test: atest TaskFragmentOrganizerControllerTest TaskTests
TaskFragmentTest
Flag: EXEMPT bugfix

Change-Id: I7486dd1aaa0d769ce638f71be66e7499c78512cd