Fixed overlap between shortcut helper header and customize buttons
in some locales due to the length of the translation of shortcut helper
header, header text was overlapping with customization buttons on the right of
shortcut helper header.

To fix this - we've added a spacer before the header text with equal
weight to the customization button row at the end of the header. we've
also limited the width of the header text to 412.dp(per uX specs) The
implications of this is that the  header text will be measured and
placed first in the row, while the spacer and the button row will get equal share
of whatever width is left after this. In the case where 412.dp isn't
enough to display the header text, it is allowed to wrap to a second
line, and truncate with ellipsis if this is still not enough.

Test: manual - using a locale with long translations ensure that
shortcut helper header text doesn't overlap with customization buttons
on the right.
Flag: com.android.systemui.keyboard_shortcut_helper_shortcut_customizer
Fix: 392763244
Fix: 387583484

Change-Id: I47c231f4ade108da21e4983928b89a21c4959879