Restore initial config and await secure setting
Some tests in InputMethodServiceTest manually modify the IME's
configuration to setup various hardware keyboard scenarios. However,
the configuration is not reset at the end of the test, and the values
persist across test runs in this class. This wraps the tests in
try-finally blocks and restore the initial config.

Aditionally, some tests rely on setting the show_ime_with_hard_keyboard
value, but don't wait for this to be applied. This adds a check that
eventually the new value takes effect. To make this more efficient, a
test API is added to IME to directly read the value of the secure
setting as cached by the IME.

Flag: EXEMPT testfix
Bug: 394328311
Test: atest InputMethodServiceTest
Change-Id: Id97df01a5603ed8ed1201da4d7ecde6ffb44650a