[Dual Shade] Introduce SceneContainerSwipeDetector for detecting swipes.
This replaces the `SplitEdgeDetector` by a more general swipe detector
for scene container, which detects swipes on the left and right halves
of the screen in addition to the edges.

This also changes the invocation gesture of the Dual Shade to not only
be swipes on the top edge of the screen, but anywhere on the screen
besides the left, right, and bottom edges.

BONUS: Removes the redundant "topEdgeSplitFraction" references and
resulting dependency on ShadeInteractor, so the swipe detector no longer
needs to be injected.

Fix: 394154602
Test: Tested manually by swiping down from both the top edge and center
 of the screen (both left and right sides) and verifying the shade opens
 and closes as expected. Also tested by swiping down when an app is open
 and verified that only swipes on the top edge expand the shade.
Test: Updated unit tests.
Flag: com.android.systemui.scene_container
Change-Id: I599bd970176c900dc44bdd1ae13d07f87a7b18bf