Notify when a fake control target changes the requested visibility
When a system bar is forcibly shown transiently, the original
controlling target will become the fake control target. And the real
control target will be InsetsPolicy#mTransientControlTarget which will
control the leash of the system bar. At this moment, the fake control
target should still be able to control the visibility of the insets
source dispatched to the client windows.

This CL calls notifyInsetsChanged when a fake control target changes the
requested visibility of its controlling type.

Bug: 393646206
Flag: EXEMPT bugfix
Test: Hide status bar in an app when a heads up notification is showing,
      and see if the app changes its layout accordingly while the
      notification it still there.
Test: ActivityRecordTests InsetsPolicyTest InsetsStateControllerTest
Change-Id: I083ac979486f1fdfa1fb50b4da0c53264624adbb