Simplify nav mode change in InputMethodServiceTest
Currently we expect an onConfigurationChanged call on the IME if
we have to change from three button to gesture navigation mode, as this
would trigger the countDownLatch too early. With the fix from [1], we
can now target the events for the countDownLatch, so this should no
longer be a concern.

This does mean that the configurationChange could be received while
handling the show request, which would cancel the ongoing request, but
the system will normally trigger a new one to continue. Submitted
separately to determine if this would introduce any race conditions in
this scenario.

  [1]: If435214366422943420a68e14ad712ce4d97fb0d

Flag: EXEMPT testfix
Bug: 394328311
Test: atest InputMethodServiceTest#testBackButtonClick
  InputMethodServiceTest#testBackButtonLongClick
  InputMethodServiceTest#testImeSwitchButtonClick
  InputMethodServiceTest#testImeSwitchButtonLongClick
Change-Id: If7b5cd23b93d1331fc16eb1458c97b76fb3ed6e7