Provide a Modifier based implementation of Expandable
This CL leverages the graphics layer/canvas retargeting API of Compose
to implement Expandable.

In the current implementation, the expanded content would be completely
moved to the overlay (in the MovableContent sense, i.e. the whole
content composition was moved) when animating.

The new implementation is now simply drawing the graphics layer of the
expandable in the overlay, making it much more closer than the View
implementation which does the same thing using GhostedView.

This new implementation is off by default and can be enabled in user
code using the useModifierBasedImplementation parameter. This will be
enabled on the new Compose Quick Settings in a follow-up CL.

Bug: 285250939
Test: Manual, set useModifierBasedImplementation to true and verified
 that all Compose Expandables were still working as expected.
Flag: EXEMPT new API not used anywhere yet
Change-Id: Iadbf2aaf32d8fb9325d5de7bd3c03b3f8eaaff3a