[Sb][Chips] Add support for composable short time delta chips
This CL adds a TimeRemainingState which can be used to track the
duration remaining between now and a given time in the future. Displays
the time remaining as "now" if less than a minute, X min if less than an
hour, or X hr if more than an hour.

Bug: b/372657935
Flag: com.android.systemui.status_bar_chips_modernization

Test: unit tests for TimeRemainingState class
Test: Make sure short time
delta chips are shown and updated screenshot in b/372657935 comment #28

Change-Id: I58eb13f7090e35c52f5a20d222973baf506f79e1