[18/N] Desks: Deactive desk on core-started task close
Deactivates the desk when handling a transition to close a task and
we're about to exit the deskt (such as when it is the last task in the
desk).

Flag: com.android.window.flags.enable_multiple_desktops_backend
Bug: 394268248
Test: open 1 app in the desk, then close with "adb shell am force-stop
<pkg-name>" to simulate core-started close. Verify we exit the desktop
properly: goes to Home, new launches launch in fullscreen and
DesksOrganizer dump shows deactivated desk.

Change-Id: I0d636306a4421c399051313660d6fb1c909100ce