Fix local coordinates for touchable region
The region and mGivenTouchableRegion are in frame-local coordinates, so
the mTmpRegion here should be in local coordinates too. As has been
pointed out in the issue tracker, the  intent here was to get the frame
size in its own coordinates instead of absolute right and bottom
coordinates.

Change-Id: I7d3256134d7e447bbd3d4e6edc937d38fe5f0ac7
Test: Not added
Flag: EXEMPT bugfix
Bug: 358274057