Setup desktop when entering desktop by app-initiated transition
This CL setup the desktop env (e.g., bring up DesktopWallpaper) when an
app is launched (or moved to top) into the desktop by app-initiated
transition when the desktop session is not activated.
We probably want to consolidate the logic with handleFreeformTaskLaunch
but this CL is the hot fix.

Flag: com.android.window.flags.enable_display_windowing_mode_switching
Bug: 397249847
Test: DesktopTasksControllerTest
Change-Id: Ia778bff76eaccadb1adb9558c4b529a145d48ea3