Update enabled for app check for Identity Check
Previously, there was only one toggle for "verify it's you" for both
face and fingerprint. As part of biometrics_onboarding_education flag
implementation, there are two separate toggles for face and fingerprint.
Identity Check needs to make sure that the toggle for the sensor that is
enrolled is enabled.

Test: Manual (See bug)
Bug: 397427773
Flag: com.android.settings.flags.biometrics_onboarding_education
Change-Id: I98f5516b0dcb3669d4f9511e7fee2f73a6c49cbb