Close idle database connections in the background
Idle database connections are now closed on the background thread
instead of the main thread.  Closing a database connection can be IO
intensive if it triggers a WAL checkpoint; do not do this on the main
thread.

Idle timeouts are deprecated as of API level 29, so this may not have
much of an effect.

Flag: EXEMPT bug-fix
Bug: 397982577
Test: atest
 * FrameworksCoreTests:android.database
 * CtsDatabaseTestCases
Change-Id: I7bd49075f978ece5b33bc491487e69c9237843df