Persist list of restored apps whose buckets need to be elevated.
Keep the list of restored apps which need to be elevated to the RARE
bucket in memory for up to 2 days (increased from 8hrs).

Also persist this list to disk in case the device restarts within the
first 2 days so important apps work as expected.

Bug: 383766428
Test: manual (ensure restore logic isn't affected)
Test: atest UsageStatsDatabaseTest
Test: atest AppIdleHistoryTests
Flag: com.android.server.usage.persist_restore_to_rare_apps_list
Change-Id: I03af37d3da8be4ce405b0abc4c0d0f79bda1574b