Don't disable the Shade's blur during launches.
This enables a closer-to-ideal launch animation for the Spatial Model,
but it does require a couple of suboptimal tricks, documented in the
code.

Ideally all of this would be much easier, but right now we can't
reliably expect all components to instantaneously update their state
correctly at the beginning and end of the animation, so we need to
work around these limitations.

Bug: 390422470
Flag: com.android.systemui.move_transition_animation_layer
Flag: com.android.systemui.notification_shade_blur
Test: manual
Change-Id: Ic80a8e6d52d82ac47e80bc53b3d10ee30c713097